Task 6 - sum

+++ 0x17-doubly_linked_lists/6-sum_dlistint.c
@@ -8,7 +8,7 @@ int sum_dlistint(dlistint_t *head)
 {
 	if (head != NULL)
 	{
-	unsigned short sum = 0;
+	int sum = 0;
 
 	while (head != NULL)
 	{
